Separate lantern adapters
