+++ src/Nethermind/Nethermind.Network.Discovery/Portal/ComponentConfiguration.cs
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
+// SPDX-License-Identifier: LGPL-3.0-only
+
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Nethermind.Network.Discovery.Portal;
+
+public class ComponentConfiguration
+{
+    public static void Configure(IServiceCollection services)
+    {
+        services.AddSingleton<IPortalContentNetworkFactory, PortalContentNetworkFactory>();
+        services.AddSingleton<IUtpManager, TalkReqUtpManager>();
+    }
+}
